+++ examples/custom_xmlrpc_client/server.py
@@ -1,7 +1,7 @@
-import time
 import random
+import time
 from SimpleXMLRPCServer import SimpleXMLRPCServer
-import xmlrpclib
+
 
 def get_time():
     time.sleep(random.random())