+++ locust/runners.py
@@ -1,20 +1,22 @@
-# coding=UTF-8
+# -*- coding: utf-8 -*-
+import logging
+import random
 import socket
 import traceback
 import warnings
-import random
-import logging
-from time import time
 from hashlib import md5
+from time import time
 
 import gevent
+import six
 from gevent import GreenletExit
 from gevent.pool import Group
 
-import events
-from stats import global_stats
+from six.moves import xrange
 
-from rpc import rpc, Message
+from . import events
+from .rpc import Message, rpc
+from .stats import global_stats
 
 logger = logging.getLogger(__name__)
 
@@ -27,6 +29,7 @@ SLAVE_REPORT_INTERVAL = 3.0
 
 class LocustRunner(object):
     def __init__(self, locust_classes, options):
+        self.options = options
         self.locust_classes = locust_classes
         self.hatch_rate = options.hatch_rate
         self.num_clients = options.num_clients
@@ -41,8 +44,9 @@ class LocustRunner(object):
         # register listener that resets stats when hatching is complete
         def on_hatch_complete(user_count):
             self.state = STATE_RUNNING
-            logger.info("Resetting stats\n")
-            self.stats.reset_all()
+            if not self.options.no_reset_stats:
+                logger.info("Resetting stats\n")
+                self.stats.reset_all()
         events.hatch_complete += on_hatch_complete
 
     @property
@@ -102,7 +106,7 @@ class LocustRunner(object):
             sleep_time = 1.0 / self.hatch_rate
             while True:
                 if not bucket:
-                    logger.info("All locusts hatched: %s" % ", ".join(["%s: %d" % (name, count) for name, count in occurence_count.iteritems()]))
+                    logger.info("All locusts hatched: %s" % ", ".join(["%s: %d" % (name, count) for name, count in six.iteritems(occurence_count)]))
                     events.hatch_complete.fire(user_count=self.num_clients)
                     return
 
@@ -225,7 +229,7 @@ class MasterLocustRunner(DistributedLocustRunner):
         
         class SlaveNodesDict(dict):
             def get_by_state(self, state):
-                return [c for c in self.itervalues() if c.state == state]
+                return [c for c in six.itervalues(self) if c.state == state]
             
             @property
             def ready(self):
@@ -260,7 +264,7 @@ class MasterLocustRunner(DistributedLocustRunner):
     
     @property
     def user_count(self):
-        return sum([c.user_count for c in self.clients.itervalues()])
+        return sum([c.user_count for c in six.itervalues(self.clients)])
     
     def start_hatching(self, locust_count, hatch_rate):
         num_slaves = len(self.clients.ready) + len(self.clients.running)
@@ -270,7 +274,7 @@ class MasterLocustRunner(DistributedLocustRunner):
             return
 
         self.num_clients = locust_count
-        slave_num_clients = locust_count / (num_slaves or 1)
+        slave_num_clients = locust_count // (num_slaves or 1)
         slave_hatch_rate = float(hatch_rate) / (num_slaves or 1)
         remaining = locust_count % num_slaves
 
@@ -281,7 +285,7 @@ class MasterLocustRunner(DistributedLocustRunner):
             self.exceptions = {}
             events.master_start_hatching.fire()
         
-        for client in self.clients.itervalues():
+        for client in six.itervalues(self.clients):
             data = {
                 "hatch_rate":slave_hatch_rate,
                 "num_clients":slave_num_clients,
@@ -305,7 +309,7 @@ class MasterLocustRunner(DistributedLocustRunner):
         events.master_stop_hatching.fire()
     
     def quit(self):
-        for client in self.clients.itervalues():
+        for client in six.itervalues(self.clients):
             self.server.send(Message("quit", None, None))
         self.greenlet.kill(block=True)
     
@@ -332,7 +336,7 @@ class MasterLocustRunner(DistributedLocustRunner):
                 self.clients[msg.node_id].state = STATE_RUNNING
                 self.clients[msg.node_id].user_count = msg.data["count"]
                 if len(self.clients.hatching) == 0:
-                    count = sum(c.user_count for c in self.clients.itervalues())
+                    count = sum(c.user_count for c in six.itervalues(self.clients))
                     events.hatch_complete.fire(user_count=count)
             elif msg.type == "quit":
                 if msg.node_id in self.clients:
@@ -348,7 +352,7 @@ class MasterLocustRunner(DistributedLocustRunner):
 class SlaveLocustRunner(DistributedLocustRunner):
     def __init__(self, *args, **kwargs):
         super(SlaveLocustRunner, self).__init__(*args, **kwargs)
-        self.client_id = socket.gethostname() + "_" + md5(str(time() + random.randint(0,10000))).hexdigest()
+        self.client_id = socket.gethostname() + "_" + md5(str(time() + random.randint(0,10000)).encode('utf-8')).hexdigest()
         
         self.client = rpc.Client(self.master_host, self.master_port)
         self.greenlet = Group()