+++ locust/test/test_stats.py
@@ -1,13 +1,14 @@
-import unittest
 import time
+import unittest
 
-from requests.exceptions import RequestException
-
-from testcases import WebserverTestCase
-from locust.stats import RequestStats, StatsEntry, global_stats
-from locust.core import HttpLocust, Locust, TaskSet, task
+from locust.core import HttpLocust, TaskSet, task
 from locust.inspectlocust import get_task_ratio_dict
 from locust.rpc.protocol import Message
+from locust.stats import RequestStats, StatsEntry, global_stats
+from six.moves import xrange
+
+from .testcases import WebserverTestCase
+
 
 class TestRequestStats(unittest.TestCase):
     def setUp(self):
@@ -97,6 +98,34 @@ class TestRequestStats(unittest.TestCase):
         self.assertEqual(s.num_failures, 3)
         self.assertEqual(s.median_response_time, 38)
         self.assertEqual(s.avg_response_time, 43.2)
+    
+    def test_error_grouping(self):
+        # reset stats
+        self.stats = RequestStats()
+        
+        s = StatsEntry(self.stats, "/some-path", "GET")
+        s.log_error(Exception("Exception!"))
+        s.log_error(Exception("Exception!"))
+            
+        self.assertEqual(1, len(self.stats.errors))
+        self.assertEqual(2, list(self.stats.errors.values())[0].occurences)
+        
+        s.log_error(Exception("Another exception!"))
+        s.log_error(Exception("Another exception!"))
+        s.log_error(Exception("Third exception!"))
+        self.assertEqual(3, len(self.stats.errors))
+    
+    def test_error_grouping_errors_with_memory_addresses(self):
+        # reset stats
+        self.stats = RequestStats()
+        class Dummy(object):
+            pass
+        
+        s = StatsEntry(self.stats, "/", "GET")
+        s.log_error(Exception("Error caused by %r" % Dummy()))
+        s.log_error(Exception("Error caused by %r" % Dummy()))
+        
+        self.assertEqual(1, len(self.stats.errors))
     
     def test_serialize_through_message(self):
         """